+++ lib/advene/gui/edit/montage.py
@@ -72,7 +72,7 @@ class Montage(AdhocView):
         # How many units does a pixel represent ?
         # self.scale.get_value() = unit by pixel
         # Unit = ms
-        self.scale = Gtk.Adjustment.new((self.controller.package.cached_duration or 60*60*1000) / Gdk.get_default_root_window().get_width(),
+        self.scale = Gtk.Adjustment.new(int((self.controller.package.cached_duration or 60*60*1000) / Gdk.get_default_root_window().get_width()),
                                         5,
                                         36000,
                                         5,
